Remove some additional enterprise license references
