--- vault/testing_util.go
@@ -1,17 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package vault
-
-import (
-	"crypto/ed25519"
-
-	testing "github.com/mitchellh/go-testing-interface"
-)
-
-func GenerateTestLicenseKeys() (ed25519.PublicKey, ed25519.PrivateKey, error) { return nil, nil, nil }
-func testExtraTestCoreSetup(testing.T, ed25519.PrivateKey, *TestClusterCore)  {}
-func testAdjustUnderlyingStorage(tcc *TestClusterCore) {
-	tcc.UnderlyingStorage = tcc.physical
-}
-func testApplyEntBaseConfig(coreConfig, base *CoreConfig) {}
