+++ website/content/docs/internals/telemetry/metrics/core-system.mdx
@@ -27,8 +27,6 @@ OpenBao instance.
 
 @include 'telemetry-metrics/vault/core/leadership_setup_failed.mdx'
 
-@include 'telemetry-metrics/vault/core/license/expiration_time_epoch.mdx'
-
 @include 'telemetry-metrics/vault/core/locked_users.mdx'
 
 @include 'telemetry-metrics/vault/core/mount_table/num_entries.mdx'
